+++ lib/mtg.dart
@@ -1,1 +1,4 @@
 library mtg;
+
+export 'util/card_parsing.dart';
+export 'util/scryfall_util.dart';